remove static methods
